v2.0.4
v2.0.4
    - Fully integrated existing functionality from v1 Gen1 SRM
    - Fully added beta support for Gen2 SRMs
        - Created CLI front end
        - Automatic detection of SRM Generation
    - Migrated defaults to it's own file [default_properies.py]
        - Changed defaults to reduce the name lengths and better describe their usage
    - Added overwrite protection
    - Updated some naming conventions to be more pythonic (and more may need to be done)
    - Added automatic game path resolution
    - Generalized srm_to_fbx to support Gen2 SRMs
    - Started adding log information
    - Minor tweak to bigfile.py allowing path to be passed as well as string
    - Renamed common.py to image.py as it only had image related functions
    - Added FbxObject and related functionality

+++ SrrConv/CLI/impls.py
@@ -2,23 +2,17 @@ from pathlib import Path
 from SrrConv.CrystalDynamics.bigfile import BigFile
 from SrrConv.CrystalDynamics.hash import hash_str
 from SrrConv import __version__
-from SrrConv.FBX.srm_gen1_to_fbx import srm_to_fbx
-from SrrConv.CrystalDynamics.common import IMAGE_FMTS
+from SrrConv.FBX.srm_to_fbx import srm_to_fbx_gen1, srm_to_fbx_gen2
+from SrrConv.CrystalDynamics.image import IMAGE_FMTS
+
+from .log import init_log
+from .default_properties import *
 
 from typing import get_args
 
 import logging
 import pickle
-
-DEFAULT_PROPERTIES = {
-    "hash_manifest": Path("./hash.manifest"),
-    "log_file": Path("./output.log"),
-    "out_path": Path("./"),
-    "image_format": "dds",
-    "sr_directory": Path("C:/Program Files (x86)/Steam/steamapps/common/Soul Reaver I-II/"),
-    "def_directory": Path("C:/Program Files (x86)/Steam/steamapps/common/Legacy of Kain Defiance Remastered/"),
-    "log_level": "info"
-}
+import tempfile
 
 def _determine_extension_from_bytes(data):
     """Determine the filetype extension utilizing the file signature"""
@@ -28,59 +22,6 @@ def _determine_extension_from_bytes(data):
         case b"SRM":
             return ".srm"
     return ".bin"
-
-def get_default_property(property, logger):
-    """Return a default property if one exists"""
-    defaults = Path("./config.defaults")
-    if not defaults.exists():
-        if logger: logger.debug("No defaults file exists...Creating defaults")
-        generate_new_defaults()
-
-    if logger: logger.debug(f"Loading {property} default from path: ./config.defaults")
-    with open("./config.defaults", "rb") as file:
-        properties = pickle.load(file)
-        if property not in properties:
-            if logger: logger.error("No property `{property}` found in defaults")
-        return properties.get(property, None)
-
-def log_level_from_str(string: str):
-    match string:
-        case "debug": return logging.DEBUG
-        case "info": return logging.INFO
-        case "warning": return logging.WARNING
-        case "error": return logging.ERROR
-        case _: return None
-
-def init_log(args):
-    """Initialize the logger using any arguments passed in the command"""
-    logger = logging.getLogger("SrrConv")
-    log_level = log_level_from_str(get_default_property("log_level", logger))
-    write_log = args.write_log
-
-    if args.log_level:
-        log_level = log_level_from_str(args.log_level)
-    
-    if log_level is None:
-        return logger
-    
-    logger.setLevel(log_level)
-
-    formatter = logging.Formatter("%(levelname)s: %(message)s")
-    stream_handler = logging.StreamHandler()
-    stream_handler.setFormatter(formatter)
-    logger.addHandler(stream_handler)
-
-    if write_log != None:
-        file_handler = logging.FileHandler(write_log, mode="w")
-        file_handler.setFormatter(formatter)
-        logger.addHandler(file_handler)
-
-    return logger
-
-def generate_new_defaults():
-    """When no config.defaults exists, create them"""
-    with open("config.defaults", "wb") as fileout:
-        pickle.dump(DEFAULT_PROPERTIES, fileout)
 
 def list_defaults_impl(args):
     """Lists the current default parameters stored inside of config.defaults"""
@@ -148,7 +89,7 @@ def hash_impl(args):
     logger.info(f"{hex(hash_str(args.string))}\t{args.string}")
 
 
-def _get_bigfile_data(big_path: str, index_type: str, input: str, logger: logging.Logger) -> None | bytearray:
+def _get_bigfile_data(big_path: str | Path, index_type: str, input: str, logger: logging.Logger) -> None | bytearray:
     """Retrieve data from a big file by hash or filename"""
     logger.info("Loading BigFile: {bigfile}")
     bigfile = BigFile.from_file(big_path)
@@ -186,19 +127,22 @@ def extract_file_impl(args):
     logger.debug(f"{args}")
 
     input = args.input
-    bigfile  = args.bigfile if args.bigfile else get_default_property("def_game_directory", logger) / "bigfilehd.dat"
+    allow_overwrite = args.allow_overwrite
+    bigfile  = args.bigfile if args.bigfile else get_default_property("def_directory", logger) / "bigfilehd.dat"
     outpath  = Path(args.outpath) if args.outpath else Path(get_default_property("out_path", logger))
     no_paths = args.no_paths
     type = args.type
 
+    final_path = _prep_output_directory(input, outpath, no_paths, logger)
+    if final_path.exists() and not allow_overwrite:
+        logger.warning("Cannot extract file: {final_path} - File exists and overwriting is disabled.")
+        return
 
     data = _get_bigfile_data(bigfile, type, input, logger)
     if data == None:
         logger.error("Failed to load data from big file! Aborting...")
         return
     logger.info("Successfully loaded data from the BigFile")
-
-    final_path = _prep_output_directory(input, outpath, no_paths, logger)
 
     logger.info(f"Writing {hex(len(data))} bytes to: {final_path}")
     with open(final_path, "wb") as file_out:
@@ -235,9 +179,10 @@ def extract_all_w_manifest_cross_ref(args):
     logger.debug(f"{args}")
 
     input    = args.input if args.input else get_default_property("hash_manifest", logger)
-    bigfile  = args.bigfile if args.bigfile else get_default_property("def_game_directory", logger) / "bigfilehd.dat"
+    bigfile  = args.bigfile if args.bigfile else get_default_property("def_directory", logger) / "bigfilehd.dat"
     outpath  = Path(args.outpath) if args.outpath else Path(get_default_property("out_path", logger))
     no_paths = args.no_paths
+    allow_overwrite = args.allow_overwrite
 
     logger.info(f"Loading BigFile: {bigfile}")
     bigfile = BigFile.from_file(bigfile)
@@ -259,6 +204,10 @@ def extract_all_w_manifest_cross_ref(args):
             name = f"{hex(hash)[2:]}{extension}"
 
         final_path = _prep_output_directory(name, outpath, no_paths, logger)
+        
+        if final_path.exists() and not allow_overwrite:
+            logger.warning("Cannot extract file: {final_path} - File exists and overwriting is disabled.")
+            continue
 
         logger.info(f"{final_path}")
         with open(final_path, "wb") as file_out:
@@ -276,6 +225,7 @@ def extract_w_manifest_cross_ref(args):
     logger.debug(f"{args}")
 
     input    = args.input
+    allow_overwrite = args.allow_overwrite
     bigfile  = args.bigfile if args.bigfile else get_default_property("def_directory", logger) / "bigfilehd.dat"
     outpath  = Path(args.outpath) if args.outpath else get_default_property("out_path", logger)
     no_paths = args.no_paths
@@ -284,9 +234,6 @@ def extract_w_manifest_cross_ref(args):
     manifest_lines = open(input, "r").readlines()
     for line in manifest_lines:
         name = line.replace('\n', '').replace('\r', '')
-        data = bigfile.get_data_from_string(name)
-        if data is None:
-            continue
 
         if no_paths:
             final_path = outpath / Path(name).name
@@ -294,6 +241,14 @@ def extract_w_manifest_cross_ref(args):
         else:
             final_path = outpath / name
             final_path.parent.mkdir(parents=True, exist_ok=True)
+
+        if final_path.exists() and not allow_overwrite:
+            logger.warning("Cannot extract file: {final_path} - File exists and overwriting is disabled.")
+            continue
+
+        data = bigfile.get_data_from_string(name)
+        if data is None:
+            continue
 
         with open(final_path, "wb") as file_out:
             file_out.write(data)
@@ -311,7 +266,7 @@ def extract_manifest_impl(args):
     else:
         return extract_w_manifest_cross_ref(args)
 
-def get_texture_dir(args, logger: logging.Logger):
+def _get_texture_dir(args, logger: logging.Logger):
     texture_dir = args.texture_dir
     if args.game_dir:
         if args.texture_dir:
@@ -321,18 +276,120 @@ def get_texture_dir(args, logger: logging.Logger):
             case "sr2": texture_dir = Path(get_default_property("sr_directory", logger)) / "2/TEX"
     return texture_dir
 
+def _get_input_dir(args, logger: logging.Logger):
+    logger.info("Resolving input directory")
+    input = Path(args.input)
+    if input.exists():
+        return input
+    
+    logger.debug("Path doesn't exist... Checking if game relative flag was set")
+    if args.game_dir:
+        match args.game_dir.lower():
+            case "sr1": 
+                logger.debug("Game relative flag SR1 set")
+                root = Path(get_default_property("sr_directory", logger)) / "1/OBJECT"
+            case "sr2": 
+                logger.debug("Game relative flag SR2 set")
+                root = Path(get_default_property("sr_directory", logger)) / "2/OBJECT"
+            case _: return None
+
+        game_relative_path = root / input
+
+        if game_relative_path.exists():
+            logger.debug(f"Path resolved: {game_relative_path}!")
+            return game_relative_path
+        else:
+            logger.error(f"Unable to resolve retalive path: {game_relative_path}")
+            return None
+    else: 
+        logger.error(f"Unable to resolve path: {input}")
+        return None
+
+def _determine_srm_version(srm_path: Path):
+    data = open(srm_path, "rb").read(4)
+    if data[-1] == 1:
+        return 1
+    if data[-1] == 8:
+        return 2
+    else:
+        return None
+    
+def _convert_file_gen1(args, logger):
+    input           = args.input
+    allow_overwrite = args.allow_overwrite
+    outpath         = args.outpath if args.outpath else (get_default_property("out_path", logger) / Path(input).stem).with_suffix(".fbx")
+    image_format    = args.image_format if args.image_format else get_default_property("image_format", logger)
+    
+    resolved_input = _get_input_dir(args, logger)
+    texture_dir = _get_texture_dir(args, logger)
+
+    srm_to_fbx_gen1(resolved_input, outpath, image_format, texture_dir, allow_overwrite, logger) # type: ignore
+    
+def _convert_file_gen2(args, logger):
+    texture_path = Path(args.texture_dir) if args.texture_dir else Path(get_default_property("def_directory", logger)) / "bigfilehd.dat"
+    input = Path(args.input)
+    outpath = Path(args.outpath) if args.outpath else (get_default_property("out_path", logger) / Path(input).stem / Path(input).stem).with_suffix(".fbx")
+    image_format    = args.image_format if args.image_format else get_default_property("image_format", logger)
+    allow_overwrite = args.allow_overwrite
+    prevent_cleanup = args.prevent_cleanup
+
+    if input.suffix.lower() != ".srm":
+        return # TODO: Error fucker
+    else:
+        srm_to_fbx_gen2(input, texture_path, outpath, image_format, allow_overwrite, prevent_cleanup, logger)
+        return
+
+    data = _get_bigfile_data(texture_path, "string", args.input, logger)
+
+    if data == None:
+        return # Error srm not found
+    
+    # Extract SRM from bigfile
+    srm_path = (outpath / Path(raw_input).stem).with_suffix(".srm")
+    with open(srm_path, "wb") as file: 
+        file.write(data)
+
+    srm_to_fbx_gen2(srm_path, texture_path, (outpath / Path(input).stem).with_suffix(".fbx"), image_format, allow_overwrite, cleanup, logger)
+
 def convert_file_impl(args):
     logger = init_log(args)
     logger.debug(f"SRRConv {__version__} - {__name__}.convert_file_impl")
     logger.debug(f"{args}")
 
-    input    = args.input
-    outpath  = args.outpath if args.outpath else (get_default_property("out_path", logger) / Path(input).stem).with_suffix(".fbx")
-    image_format  = args.image_format if args.image_format else get_default_property("image_format", logger)
+    resolved_input = _get_input_dir(args, logger)
+    if resolved_input == None:
+        return # File not found
     
-    texture_dir = get_texture_dir(args, logger)
+    if resolved_input.suffix.lower() != ".srm":
+        return # error must be srm
+    version = _determine_srm_version(resolved_input)
     
-    if texture_dir is None:
-        return # failure
+    if version == 1:
+        return _convert_file_gen1(args, logger)
+    else:
+        return _convert_file_gen2(args, logger)
     
-    srm_to_fbx(input, outpath, image_format, texture_dir, logger)
+    # allow_overwrite = args.allow_overwrite
+    # outpath         = args.outpath if args.outpath else get_default_property("out_path", logger) / Path(raw_input).stem
+    # image_format    = args.image_format if args.image_format else get_default_property("image_format", logger)
+
+    # if version == 1:
+    #     texture_dir = _get_texture_dir(args, logger)
+    #     srm_to_fbx_gen1(resolved_input, (outpath / Path(raw_input).stem).with_suffix(".fbx"), image_format, texture_dir, allow_overwrite, logger)
+    # else:
+    #     # If it's already a srm file do the thing
+    #     if resolved_input.suffix.lower() == ".srm":
+    #         srm_to_fbx_gen2(resolved_input, args.input, (outpath / Path(raw_input).stem).with_suffix(".fbx"), image_format, allow_overwrite, logger)
+    #         return
+
+    #     data = _get_bigfile_data(resolved_input, "string", args.input, logger)
+
+    #     if data == None:
+    #         return # Error srm not found
+        
+    #     # Extract SRM from bigfile
+    #     srm_path = (outpath / Path(raw_input).stem).with_suffix(".srm")
+    #     with open(srm_path, "wb") as file: 
+    #         file.write(data)
+
+    #     srm_to_fbx_gen2(resolved_input, args.input, (outpath / Path(raw_input).stem).with_suffix(".fbx"), image_format, allow_overwrite, logger)
